Backend/requirements.txt
Flask
requests
assemblyai
python-dotenv
transformers
scipy
moviepy